Domain layer was decoupled and Sales feature added
