+++ app/build.gradle
@@ -58,10 +58,11 @@ dependencies {
     implementation "androidx.lifecycle:lifecycle-viewmodel-compose:2.6.0-alpha04"
     implementation 'androidx.compose.material3:material3:1.1.0-alpha04'
     implementation 'androidx.compose.material:material:1.3.1'
-    implementation project(path: ':demosales-domain')
+//    implementation project(path: ':demosales-domain')
     implementation project(path: ':demosales-data:products')
     implementation project(path: ':demosales-ui:products')
     implementation project(path: ':demosales-datasource:room')
+    implementation project(path: ':demosales-domain:products')
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
