+++ demosales-datasource/room/build.gradle
@@ -36,8 +36,9 @@ dependencies {
     implementation 'androidx.core:core-ktx:1.9.0'
     implementation 'androidx.appcompat:appcompat:1.6.0'
     implementation project(path: ':demosales-data:products')
-    implementation project(path: ':demosales-domain')
+//    implementation project(path: ':demosales-domain')
     implementation 'androidx.test.ext:junit-ktx:1.1.5'
+    implementation project(path: ':demosales-domain:products')
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
