+++ demosales-domain/products/src/main/java/com/imecatro/demosales/domain/products/products/model/Currencies.kt
@@ -1,0 +1,8 @@
+package com.imecatro.demosales.domain.products.products.model
+
+sealed class Currencies(val code: String) {
+    object Dollar : Currencies("USD")
+    object MexicanPeso : Currencies("MXN")
+    object Euros : Currencies("EUR")
+}
+
