+++ demosales-domain/products/src/main/java/com/imecatro/demosales/domain/products/products/repository/ProductsRepository.kt
@@ -1,0 +1,12 @@
+package com.imecatro.demosales.domain.products.products.repository
+
+import com.imecatro.demosales.domain.products.products.model.ProductDomainModel
+import kotlinx.coroutines.flow.Flow
+
+interface ProductsRepository {
+    fun addProduct(product: ProductDomainModel?)
+     fun getAllProducts(): Flow<List<ProductDomainModel>>
+    fun deleteProductById(id :Int?)
+    fun updateProduct(product: ProductDomainModel?)
+    fun getProductDetailsById(id: Int?): ProductDomainModel?
+}
