+++ demosales-domain/products/src/main/java/com/imecatro/demosales/domain/products/products/usecases/DeleteProductByIdUseCase.kt
@@ -1,0 +1,14 @@
+package com.imecatro.demosales.domain.products.products.usecases
+
+import com.imecatro.demosales.domain.products.products.repository.ProductsRepository
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.Dispatchers
+
+class DeleteProductByIdUseCase(
+    private val productsRepository: ProductsRepository,
+    private val dispatcher: CoroutineDispatcher = Dispatchers.IO
+) {
+    operator fun invoke(id : Int?) {
+        productsRepository.deleteProductById(id)
+    }
+}
