+++ demosales-domain/products/src/main/java/com/imecatro/demosales/domain/products/products/usecases/GetAllProductsUseCase.kt
@@ -1,0 +1,18 @@
+package com.imecatro.demosales.domain.products.products.usecases
+
+import com.imecatro.demosales.domain.products.products.model.ProductDomainModel
+import com.imecatro.demosales.domain.products.products.repository.ProductsRepository
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.withContext
+
+class GetAllProductsUseCase(
+    private val productsRepository: ProductsRepository,
+    private val dispatcher: CoroutineDispatcher = Dispatchers.IO
+) {
+    suspend operator fun invoke(): Flow<List<ProductDomainModel>> =
+        withContext(dispatcher) {
+            return@withContext productsRepository.getAllProducts()
+        }
+}
