--- demosales-domain/src/main/java/com/imecatro/domain/products/model/Currencies.kt
@@ -1,8 +1,0 @@
-package com.imecatro.domain.products.model
-
-sealed class Currencies(val code: String) {
-    object Dollar : Currencies("USD")
-    object MexicanPeso : Currencies("MXN")
-    object Euros : Currencies("EUR")
-}
-
