--- demosales-domain/src/main/java/com/imecatro/domain/products/model/ProductDomainModel.kt
@@ -1,13 +1,0 @@
-package com.imecatro.domain.products.model
-
-
-data class ProductDomainModel(
-    var id: Int?,
-    val name: String?,
-    val price: Float?,
-    val currency: String?,
-    val unit: ProductUnit = ProductUnit.Default,
-    val details: String,
-    val imageUri : String?
-)
-
