--- demosales-domain/src/main/java/com/imecatro/domain/products/usecases/UpdateProductUseCase.kt
@@ -1,11 +1,0 @@
-package com.imecatro.domain.products.usecases
-
-import com.imecatro.domain.products.model.ProductDomainModel
-import com.imecatro.domain.products.repository.ProductsRepository
-
-class UpdateProductUseCase(private val productsRepository: ProductsRepository) {
-    operator fun invoke(product: ProductDomainModel) {
-        productsRepository.updateProduct(product)
-    }
-
-}
