+++ demosales-ui/products/build.gradle
@@ -44,7 +44,8 @@ dependencies {
     implementation 'androidx.core:core-ktx:1.9.0'
     implementation 'androidx.appcompat:appcompat:1.6.0'
     implementation 'com.google.android.material:material:1.7.0'
-    implementation project(path: ':demosales-domain')
+    implementation project(path: ':demosales-domain:products')
+//    implementation project(path: ':demosales-domain')
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
