+++ demosales-ui/products/src/main/java/com/imecatro/products/ui/add/mappers/AddProductsUiMappers.kt
@@ -1,10 +1,10 @@
 package com.imecatro.products.ui.add.mappers
 
-import com.imecatro.domain.products.model.ProductDomainModel
-import com.imecatro.domain.products.model.ProductUnit
+import com.imecatro.demosales.domain.products.products.model.ProductDomainModel
+import com.imecatro.demosales.domain.products.products.model.ProductUnit
 import com.imecatro.products.ui.add.model.AddProductUiModel
 
- fun AddProductUiModel.toDomain(): ProductDomainModel? {
+fun AddProductUiModel.toDomain(): ProductDomainModel? {
     return ProductDomainModel(
         id = null,
         name = this.name,
