+++ demosales-ui/products/src/main/java/com/imecatro/products/ui/details/mappers/ProductsDetailsUiMappers.kt
@@ -1,6 +1,6 @@
 package com.imecatro.products.ui.details.mappers
 
-import com.imecatro.domain.products.model.ProductDomainModel
+import com.imecatro.demosales.domain.products.products.model.ProductDomainModel
 import com.imecatro.products.ui.details.model.ProductDetailsUiModel
 
 fun ProductDomainModel.toUiModel(): ProductDetailsUiModel {
@@ -8,6 +8,7 @@ fun ProductDomainModel.toUiModel(): ProductDetailsUiModel {
         id = this.id,
         name = this.name,
         price = this.price.toString(),
+        currency = this.currency,
         unit = this.unit.symbol,
         imageUrl = this.imageUri,
         details = this.details
