+++ demosales-ui/products/src/main/java/com/imecatro/products/ui/details/viewmodels/ProductsDetailsViewModel.kt
@@ -2,7 +2,7 @@ package com.imecatro.products.ui.details.viewmodels
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.imecatro.domain.products.repository.ProductsRepository
+import com.imecatro.demosales.domain.products.products.repository.ProductsRepository
 import com.imecatro.products.ui.details.mappers.toUiModel
 import com.imecatro.products.ui.details.model.ProductDetailsUiModel
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -18,14 +18,14 @@ class ProductsDetailsViewModel @Inject constructor(
     private val productsRepository: ProductsRepository //= ProductsRepositoryDummyImpl()
 ) : ViewModel() {
 
-    private val initDetails = ProductDetailsUiModel(0, "Product", "0.00", "pz", null, "")
+    private val initDetails = ProductDetailsUiModel(0, "Product", "0.00", "", "", null, "")
 
     private val _product: MutableStateFlow<ProductDetailsUiModel?> = MutableStateFlow(initDetails)
     val product: StateFlow<ProductDetailsUiModel?> = _product.asStateFlow()
 
     fun getDetailsById(id: Int?) {
         viewModelScope.launch(Dispatchers.IO) {
-            val response  = productsRepository.getProductDetailsById(id)?.toUiModel()
+            val response = productsRepository.getProductDetailsById(id)?.toUiModel()
             _product.value = response
         }
