+++ demosales-ui/products/src/main/java/com/imecatro/products/ui/details/views/BottomSheetDetailsComposable.kt
@@ -2,6 +2,7 @@ package com.imecatro.products.ui.details.views
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.material.*
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Delete
@@ -43,7 +44,7 @@ fun DetailsComposableStateImpl(
         onDeleteClicked = {
             onNavigateToEdit(null)
             productDetailsViewModel.onDeleteAction(productId)
-                          },
+        },
         onEditClicked = {
             productId?.let {
                 onNavigateToEdit(it)
@@ -59,66 +60,73 @@ fun DetailsComposable(
     onEditClicked: () -> Unit
 ) {
     val paddingX = 20.dp
-    Column(modifier = Modifier.fillMaxSize(), horizontalAlignment = Alignment.CenterHorizontally) {
-
-        Spacer(modifier = Modifier.height(20.dp))
-        Image(
-            painter = rememberAsyncImagePainter(
-                //productDetails?.imageUrl ?: R.raw.arcreactor
-                ImageRequest.Builder(LocalContext.current)
-                    .data(productDetails?.imageUrl)
-                    .error(R.drawable.baseline_insert_photo_24)
-                    .crossfade(true)
-                    .build()
-            ),
-            contentDescription = null,
-            modifier = Modifier
-                .fillMaxWidth()
-                .requiredSizeIn(maxHeight = 280.dp)
+    LazyColumn(
+        modifier = Modifier.fillMaxSize(),
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        item {
+            Spacer(modifier = Modifier.height(20.dp))
+            Image(
+                painter = rememberAsyncImagePainter(
+                    //productDetails?.imageUrl ?: R.raw.arcreactor
+                    ImageRequest.Builder(LocalContext.current)
+                        .data(productDetails?.imageUrl)
+                        .error(R.drawable.baseline_insert_photo_24)
+                        .crossfade(true)
+                        .build()
+                ),
+                contentDescription = null,
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .requiredSizeIn(maxHeight = 280.dp)
 //                .padding(paddingX)
 //                .clip(RoundedCornerShape(10))
-            ,
-            contentScale = ContentScale.FillWidth
-        )
-        Spacer(modifier = Modifier.height(20.dp))
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(20.dp, 10.dp),
-            horizontalAlignment = Alignment.Start
-        ) {
-            //tittle
-            Text(text = productDetails?.name ?: "No name", style = Typography.titleMedium)
-            //pz
-            Text(text = productDetails?.unit ?: "", style = Typography.bodyLarge)
-            //price
-            Text(text =  "$${productDetails?.price ?: "0.00"}", style = Typography.titleMedium)
+                ,
+                contentScale = ContentScale.FillWidth
+            )
             Spacer(modifier = Modifier.height(20.dp))
-            //Details
-            Text(text = "Details", style = Typography.labelMedium, color = PurpleGrey40)
-            Divider(color = Color.LightGray, thickness = 2.dp)
-            Text(
-                text = productDetails?.details
-                    ?: "This text must contains some details about this product",
-                style = Typography.bodyLarge
+            Column(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(20.dp, 10.dp),
+                horizontalAlignment = Alignment.Start
+            ) {
+                //tittle
+                Text(text = productDetails?.name ?: "No name", style = Typography.titleMedium)
+                //pz
+                Text(text = productDetails?.unit ?: "", style = Typography.bodyLarge)
+                //price
+                Text(
+                    text = "$${productDetails?.price ?: "0.00"} ${productDetails?.currency ?: ""}",
+                    style = Typography.titleMedium
+                )
+                Spacer(modifier = Modifier.height(20.dp))
+                //Details
+                Text(text = "Details", style = Typography.labelMedium, color = PurpleGrey40)
+                Divider(color = Color.LightGray, thickness = 2.dp)
+                Text(
+                    text = productDetails?.details
+                        ?: "This text must contains some details about this product",
+                    style = Typography.bodyLarge
+                )
+            }
+            Spacer(modifier = Modifier.height(80.dp))
+            //button edit
+            ButtonFancy(
+                text = "EDIT",
+                paddingX = paddingX,
+                icon = Icons.Filled.Edit,
+                onClicked = onEditClicked
+            )
+            //button delete
+            ButtonFancy(
+                text = "DELETE",
+                color = PurpleRed,
+                paddingX = paddingX,
+                icon = Icons.Filled.Delete,
+                onClicked = onDeleteClicked
             )
         }
-        Spacer(modifier = Modifier.height(80.dp))
-        //button edit
-        ButtonFancy(
-            text = "EDIT",
-            paddingX = paddingX,
-            icon = Icons.Filled.Edit,
-            onClicked = onEditClicked
-        )
-        //button delete
-        ButtonFancy(
-            text = "DELETE",
-            color = PurpleRed,
-            paddingX = paddingX,
-            icon = Icons.Filled.Delete,
-            onClicked = onDeleteClicked
-        )
     }
 }
 
@@ -136,6 +144,7 @@ fun PreviewWordsListDetailsCompose() {
                     "Cebolla",
                     "$0.00",
                     "x pz",
+                    "USD",
                     null,
                     "details"
                 ),
