+++ demosales-ui/products/src/main/java/com/imecatro/products/ui/update/mappers/UpdateProductUiMappers.kt
@@ -1,8 +1,8 @@
 package com.imecatro.products.ui.update.mappers
 
 import android.net.Uri
-import com.imecatro.domain.products.model.ProductDomainModel
-import com.imecatro.domain.products.model.ProductUnit
+import com.imecatro.demosales.domain.products.products.model.ProductDomainModel
+import com.imecatro.demosales.domain.products.products.model.ProductUnit
 import com.imecatro.products.ui.update.model.UpdateProductUiModel
 
 
@@ -12,13 +12,14 @@ fun UpdateProductUiModel.toDomain(): ProductDomainModel {
         name = this.name,
         price = this.price?.toFloat() ?: 0f,
         currency = this.currency,
-        unit = ProductUnit::class.nestedClasses.map { it.objectInstance as ProductUnit }.find { it.symbol == this.unit }?: ProductUnit.Default,
+        unit = ProductUnit::class.nestedClasses.map { it.objectInstance as ProductUnit }
+            .find { it.symbol == this.unit } ?: ProductUnit.Default,
         details = this.details,
         imageUri = this.imageUri?.toString()
     )
 }
 
-fun ProductDomainModel.toUpdateUiModel() : UpdateProductUiModel{
+fun ProductDomainModel.toUpdateUiModel(): UpdateProductUiModel {
     return UpdateProductUiModel(
         id = this.id,
         name = this.name,
@@ -26,6 +27,6 @@ fun ProductDomainModel.toUpdateUiModel() : UpdateProductUiModel{
         currency = this.currency,
         unit = this.unit.symbol,
         details = this.details,
-        imageUri = Uri.parse( this.imageUri.toString())
+        imageUri = Uri.parse(this.imageUri.toString())
     )
 }
