+++ demosales-ui/products/src/main/java/com/imecatro/products/ui/update/views/UpdateProductComposable.kt
@@ -84,8 +84,10 @@ fun UpdateProductComposableStateImpl(
         productPrice = productPrice ?: "0.00",
         onProductPriceChange = { productPrice = it },
         currencyList = updateProductViewModel.getCurrencies(),
+        currencyPicked = currencySelected?: "USD",
         onCurrencyChange = { currencySelected = it },
         unitList = updateProductViewModel.getUnities(),
+        unitPicked = unitSelected ?: "pz",
         onUnitPicked = { unitSelected = it },
         detailsText = details ?: "",
         onDetailsChange = { details = it },
