+++ settings.gradle
@@ -14,7 +14,8 @@ dependencyResolutionManagement {
 }
 rootProject.name = "PuntroSales-Demo"
 include ':app'
-include ':demosales-domain'
 include ':demosales-data:products'
 include ':demosales-ui:products'
 include ':demosales-datasource:room'
+include ':demosales-domain:sales'
+include ':demosales-domain:products'
